use inheritance from `OpenTracing` base classes
Use proper inheritance from the `OpenTracing` base classes for `Span` and `Tracer` as indicated here: https://github.com/opentracing/opentracing-javascript#opentracing-tracer-implementations

The major change is the usage of `TraceId` as a `span context` which is returned from `Tracer.extract()` instead of a `span`. The unit tests have been adjusted accordingly. And the implementation have been simplified by this.

Since I don't know much about the `zipkin` protocol itself, I was not able to fill in some reasonable implementations for the `baggage` feature of `OpenTracing`.